requirements.txt
requests
python-dotenv
bsu
fastapi
selenium
uvicorn

